# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Molecular Vision AR (MVAR) API Documentation",
        "description": "Dokumentasi RESTful API untuk Molecular Vision AR (MVAR) Backend. API ini digunakan oleh aplikasi mobile (Flutter / Unity) dan client lainnya.",
        "contact": {
            "name": "Tim Pengembang MVAR",
            "email": "dev@mvar.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Ambil Profil Pengguna (Me)",
                "description": "Mengembalikan detail profil pengguna yang sedang login berdasarkan Bearer Token Sanctum.",
                "operationId": "bc2032a379aa97707ccc61511897c120",
                "responses": {
                    "200": {
                        "description": "Profil pengguna berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data profil berhasil diambil."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id_user": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Budi Santoso"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "budi@example.com"
                                                        },
                                                        "school": {
                                                            "type": "string",
                                                            "example": "SMAN 1 Jakarta",
                                                            "nullable": true
                                                        },
                                                        "id_degree": {
                                                            "type": "integer",
                                                            "example": 1,
                                                            "nullable": true
                                                        },
                                                        "id_role": {
                                                            "type": "integer",
                                                            "example": 2,
                                                            "nullable": true
                                                        },
                                                        "role": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "siswa (mobile)"
                                                                }
                                                            },
                                                            "type": "object",
                                                            "nullable": true
                                                        },
                                                        "degree": {
                                                            "properties": {
                                                                "id_degree": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "SMA / MA"
                                                                }
                                                            },
                                                            "type": "object",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau tidak disertakan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout Pengguna (Cabut Token)",
                "description": "Mencabut (menghapus) token personal access Sanctum yang sedang aktif digunakan.",
                "operationId": "091884718461f829a9a6277f90c8be66",
                "responses": {
                    "200": {
                        "description": "Logout berhasil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil. Token telah dicabut."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau sudah tidak aktif.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token Sanctum dengan format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}